Introduce leaf_proxy_map::as_set().
In support of which, make leaf_proxy objects comparable by sorting against key.
Add at test that fails without this change but that succeeds with it.
Fix some other tests that were passing by coincidence and that are now deterministic.
In root_test.cpp, move static functions to anonymous namespace.

Closes #19.