Fix some const-correctness problems in leaf_proxy.
Make the_leaf a mutable pointer, since loading the leaf doesn't
conceptually change the state of the leaf_proxy.  Don't cache quite so
much in the leaf_proxy: if the leaf is authoritative, let it be.  Make
some functions const that couldn't be const before.  Conversely, don't
return const string's, this was a bit pointless.

Fix a bug where we didn't always commit leaves promptly.

Closes #22.