Add delete command to srd.
Refactor do_match to use a visitor (print or delete).
Add help text for deletion.
Improve help text for import and create.
Fix a bug in which the root was not committed on leaf erase.
Fix a bug in which the root was not committed promptly on leaf add.
Fix some const correctness points.

Cf. #16.