Fix key change bug.
Fix bug in which changing a key value didn't flag the root as
modified, and so the root would not be rewritten, and would appear
invalid on next load due to the key disagreeing with the cached key.

Add a test for the fix.  Confirmed that the test fails without this
fix.